move browser detection one level up to root component
